chore(themes): generate extended bootstrap source
