+++ oxfmt.config.ts
@@ -24,5 +24,6 @@ export default defineConfig({
 		"pnpm-lock.yaml",
 		// Generated by packages/themes/scripts/generate-script-source.ts
 		"packages/themes/src/core/script-source.ts",
+		"packages/themes/src/core/extended-script-source.ts",
 	],
 } satisfies OxfmtConfig);
