+++ packages/themes/scripts/generate-script-source.ts
@@ -2,9 +2,28 @@ import { mkdir, readFile, writeFile } from "node:fs/promises";
 import { dirname, resolve } from "node:path";
 
 const rootDir = resolve(import.meta.dir, "..");
-const bootstrapPath = resolve(rootDir, "src/core/script-bootstrap.ts");
-const outputPath = resolve(rootDir, "src/core/script-source.ts");
 const checkOnly = process.argv.includes("--check");
+
+const targets = [
+	{
+		bootstrapPath: resolve(rootDir, "src/core/script-bootstrap.ts"),
+		outputPath: resolve(rootDir, "src/core/script-source.ts"),
+		displayPath: "src/core/script-source.ts",
+		readableName: "script-bootstrap.ts",
+		exportName: "THEME_SCRIPT_SOURCE",
+		label: "theme bootstrap",
+	},
+	{
+		bootstrapPath: resolve(rootDir, "src/core/extended-script-bootstrap.ts"),
+		outputPath: resolve(rootDir, "src/core/extended-script-source.ts"),
+		displayPath: "src/core/extended-script-source.ts",
+		readableName: "extended-script-bootstrap.ts",
+		exportName: "EXTENDED_THEME_SCRIPT_SOURCE",
+		label: "extended theme bootstrap",
+	},
+] as const;
+
+type SourceTarget = (typeof targets)[number];
 
 function extractMinifiedFunction(bundle: string): string {
 	const start = bundle.search(/function\s+\w+\s*\(/u);
@@ -83,15 +102,15 @@ function extractMinifiedFunction(bundle: string): string {
 	throw new Error("Failed to find the end of the minified bootstrap function");
 }
 
-function renderSourceFile(minified: string): string {
-	return `/** Generated by scripts/generate-script-source.ts — edit script-bootstrap.ts instead. */
-export const THEME_SCRIPT_SOURCE: string = ${JSON.stringify(minified)};
+function renderSourceFile(target: SourceTarget, minified: string): string {
+	return `/** Generated by scripts/generate-script-source.ts — edit ${target.readableName} instead. */
+export const ${target.exportName}: string = ${JSON.stringify(minified)};
 `;
 }
 
-async function minifyBootstrap(): Promise<string> {
+async function minifyBootstrap(target: SourceTarget): Promise<string> {
 	const result = await Bun.build({
-		entrypoints: [bootstrapPath],
+		entrypoints: [target.bootstrapPath],
 		minify: true,
 		target: "browser",
 		write: false,
@@ -99,39 +118,49 @@ async function minifyBootstrap(): Promise<string> {
 
 	if (!result.success) {
 		const messages = result.logs.map((log) => log.message ?? String(log)).join("\n");
-		throw new Error(`Failed to minify theme bootstrap:\n${messages}`);
+		throw new Error(`Failed to minify ${target.label}:\n${messages}`);
 	}
 
 	const output = result.outputs[0];
 	if (!output) {
-		throw new Error("Bun.build produced no minify output for theme bootstrap");
+		throw new Error(`Bun.build produced no minify output for ${target.label}`);
 	}
 
 	return extractMinifiedFunction(await output.text());
 }
 
-const minified = await minifyBootstrap();
-const nextContents = renderSourceFile(minified);
+const generated = await Promise.all(
+	targets.map(async (target) => ({
+		target,
+		contents: renderSourceFile(target, await minifyBootstrap(target)),
+	})),
+);
 
 if (checkOnly) {
-	let current = "";
-	try {
-		current = await readFile(outputPath, "utf8");
-	} catch {
-		current = "";
+	const stalePaths: string[] = [];
+	for (const { target, contents } of generated) {
+		let current = "";
+		try {
+			current = await readFile(target.outputPath, "utf8");
+		} catch {
+			current = "";
+		}
+		if (current !== contents) stalePaths.push(target.displayPath);
 	}
 
-	if (current !== nextContents) {
+	if (stalePaths.length > 0) {
 		console.error(
-			"src/core/script-source.ts is out of date. Run `bun scripts/generate-script-source.ts`.",
+			`${stalePaths.join(", ")} ${stalePaths.length === 1 ? "is" : "are"} out of date. Run \`bun scripts/generate-script-source.ts\`.`,
 		);
 		process.exit(1);
 	}
 
-	console.log("src/core/script-source.ts is up to date.");
+	console.log("Generated script sources are up to date.");
 	process.exit(0);
 }
 
-await mkdir(dirname(outputPath), { recursive: true });
-await writeFile(outputPath, nextContents);
-console.log(`Wrote ${outputPath}`);
+for (const { target, contents } of generated) {
+	await mkdir(dirname(target.outputPath), { recursive: true });
+	await writeFile(target.outputPath, contents);
+	console.log(`Wrote ${target.outputPath}`);
+}
