+++ packages/themes/src/core/extended-script-source.ts
@@ -1,0 +1,2 @@
+/** Generated by scripts/generate-script-source.ts — edit extended-script-bootstrap.ts instead. */
+export const EXTENDED_THEME_SCRIPT_SOURCE: string = "function(u,g,A,m,q,v,f,R,p,i,c,a,h,w,y){let s;if(v)s=v;else if(a&&(f.includes(a)||m&&a===\"system\"))s=a;else{let t=null;if(!w)try{if(i===\"cookie\"||i===\"hybrid\"){let e=(\"; \"+document.cookie).split(\"; \"+u+\"=\"),S=e.length>1?e.pop()?.split(\";\")[0]:null,L=S?decodeURIComponent(S):null;t=i===\"hybrid\"?L??localStorage.getItem(u):L}else if(i!==\"none\")t=(i===\"localStorage\"?localStorage:sessionStorage).getItem(u)}catch{}s=t&&(f.includes(t)||m&&t===\"system\")?t:A}let d=m?matchMedia(\"(prefers-color-scheme: dark)\").matches?\"dark\":\"light\":null,n=s;if(d){let t=y;if(t&&typeof t.light===\"string\"&&typeof t.dark===\"string\"){if(s===\"system\")n=t[d]}else if(y&&s!==\"system\")n=y[s]?.[d]??s;else if(s===\"system\")n=d}else if(s===\"system\")n=A;let r=R?.[n]||n,o=p===\"html\"?document.documentElement:p===\"body\"?document.body:document.querySelector(p);if(!o)return;let x=Array.isArray(g)?g:[g],E=f.flatMap((t)=>(R?.[t]||t).split(\" \")),k=r.split(\" \"),l=!1,b=!1;for(let t of x)if(t===\"class\")b=E.some((e)=>!k.includes(e)&&o.classList.contains(e))||k.some((e)=>!o.classList.contains(e)),l=l||b;else l=l||o.getAttribute(t)!==r;if(l&&h){let t=typeof h===\"string\"?h:\"none\",e=document.createElement(\"style\");e.textContent=\"*,*::before,*::after{transition:\"+t+\"!important}\",document.head.appendChild(e),requestAnimationFrame(()=>requestAnimationFrame(()=>e.remove()))}for(let t of x)if(t===\"class\"){if(b)o.classList.remove(...E),o.classList.add(...k)}else if(r){if(o.getAttribute(t)!==r)o.setAttribute(t,r)}else o.removeAttribute(t);if(q&&(n===\"light\"||n===\"dark\"))o.style.colorScheme=n;if(c){let t=typeof c===\"string\"?c:c[n];if(t){let e=document.querySelector('meta[name=\"theme-color\"]');if(!e)e=document.createElement(\"meta\"),e.setAttribute(\"name\",\"theme-color\"),document.head.appendChild(e);e.setAttribute(\"content\",t)}}}";
